Reduction of the size of the filter instance by removing initialization state from the class members.
Initialization of the filter (X/Y/Z and gyro/acco) in the calibration
procedure with the current mesurement in order to improve and reduce the
convergence.

Add of the estimation of produced torsor (torque and force) in the
modulator. Force can be used by other function as for instance future
contact force detection function which will be used to detect landed
state, but also colisions or so on.